huge update with folder architecture and changes with @sinaendhran
