+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "cSpell.words": [
+        "aerostructural",
+        "TACS"
+    ]
+}
