+++ examples/inputs/Local/naca0012_simInfo.yaml
@@ -1,32 +1,32 @@
 out_dir: output
-hpc: 'no'
-nproc: 4
+machine_type: local
+nproc: 8
 hierarchies:
 - name: 2d_clean
   cases:
   - name: NACA0012
     problem: Aero
-    meshes_folder_path: ./grids/
+    meshes_folder_path: /home/sanjan/PhD/mdss/examples/grids
     mesh_files:
     - naca0012_L0.cgns
     geometry_info:
       chordRef: 1.0
       areaRef: 1.0
     aero_options:
-      printIterations: true
       useANKSolver: true
-      nSubiterTurb: 10
+      nSubiterTurb: 20
       useNKSolver: false
-      ANKCoupledSwitchTol: 1.0e-3
-      ANKSecondOrdSwitchTol: 1.0e-8
+      NKSwitchTol: 1.0e-06
+      ANKCoupledSwitchTol: 0.001
+      ANKSecondOrdSwitchTol: 1.0e-12
       L2Convergence: 1.0e-08
       nCycles: 150000
     scenarios:
     - name: cruise_1
       aoa_list:
-      - 0
-      - 5
-      Re: 1000000.0
-      mach: 0.7
-      Temp: 298.0
-      exp_data: exp_data/naca0012.txt
+        - 0
+        - 5
+      reynolds: 3900000.0
+      mach: 0.3
+      T: 298.0
+      exp_data: /home/sanjan/PhD/mdss/examples/exp_data/naca0012.csv
