+++ mdss/__init__.py
@@ -1,0 +1,9 @@
+__version__ = "0.1.0"
+
+from .src import main
+from .utils import utils
+from .src.aerostruct import Problem
+from .src.main import simulation
+from .src.main_helper import execute
+
+__all__ = ["main", "utils", "Problem", "simulation", "execute"]
