--- mdss/resources/30p-30n_simInfo.yaml
@@ -1,36 +1,0 @@
-out_dir: output
-hpc: 'no'
-nproc: 8 
-hierarchies:
-- name: 2d_high_lift
-  cases:
-  - name: 30p-30n
-    problem: Aero
-    meshes_folder_path: ../grids
-    mesh_files:
-    - 30p-30n_overset_L0.cgns
-    - 30p-30n_overset_L1.cgns
-    - 30p-30n_overset_L2.cgns
-    geometry_info:
-      chordRef: 1.0
-      areaRef: 1.0
-    solver_parameters:
-      useANKSolver: true
-      nSubiterTurb: 20
-      useNKSolver: false
-      NKSwitchTol: 0.0001
-      ANKCoupledSwitchTol: 1.0e-07
-      ANKSecondOrdSwitchTol: 1.0e-05
-      L2Convergence: 1.0e-10
-      nCycles: 150000
-      liftIndex: 2
-      nearWallDist: 0.01
-    scenarios:
-    - name: cruise_1
-      aoa_list:
-        - 5
-        - 10
-      Re: 9000000.0
-      mach: 0.2
-      Temp: 298.0
-      exp_data: ./exp_data/30p-30n.txt
