+++ mdss/resources/aero_defaults.py
@@ -1,0 +1,40 @@
+# This python file stores the defaults
+
+################################################################################
+# Default Adflow solver options for Aerodynamic problem
+################################################################################
+default_aero_options_aerodynamic = {
+    # Print Options
+    "printIterations": False,
+    "printAllOptions": False,
+    "printIntro": False,
+    "printTiming": False,
+    # I/O Parameters
+    "gridFile": f"grids/naca0012_L1.cgns", # Default grid file
+    "outputDirectory": ".",
+    "monitorvariables": ["resrho", "resturb", "cl", "cd", "yplus"],
+    "writeTecplotSurfaceSolution": True,
+    # Physics Parameters
+    "equationType": "RANS",
+    "liftindex": 3,  # z is the lift direction
+    # Solver Parameters
+    "smoother": "DADI",
+    "CFL": 0.5,
+    "CFLCoarse": 0.25,
+    "MGCycle": "sg",
+    "MGStartLevel": -1,
+    "nCyclesCoarse": 250,
+    # ANK Solver Parameters
+    "useANKSolver": True,
+    "nSubiterTurb": 5,
+    "ANKSecondOrdSwitchTol": 1e-4,
+    "ANKCoupledSwitchTol": 1e-6,
+    "ankinnerpreconits": 2,
+    "ankouterpreconits": 2,
+    "anklinresmax": 0.1,
+    # Termination Criteria
+    "L2Convergence": 1e-12,
+    "L2ConvergenceCoarse": 1e-2,
+    "nCycles": 75000,
+}
+
