+++ mdss/resources/aerostruct_defaults.py
@@ -1,43 +1,3 @@
-# This python file stores the defaults
-
-################################################################################
-# Default Adflow solver options for Aerodynamic problem
-################################################################################
-default_aero_options_aerodynamic = {
-    # Print Options
-    "printIterations": False,
-    "printAllOptions": False,
-    "printIntro": False,
-    "printTiming": False,
-    # I/O Parameters
-    "gridFile": f"grids/naca0012_L1.cgns", # Default grid file
-    "outputDirectory": ".",
-    "monitorvariables": ["resrho", "resturb", "cl", "cd", "yplus"],
-    "writeTecplotSurfaceSolution": True,
-    # Physics Parameters
-    "equationType": "RANS",
-    "liftindex": 3,  # z is the lift direction
-    # Solver Parameters
-    "smoother": "DADI",
-    "CFL": 0.5,
-    "CFLCoarse": 0.25,
-    "MGCycle": "sg",
-    "MGStartLevel": -1,
-    "nCyclesCoarse": 250,
-    # ANK Solver Parameters
-    "useANKSolver": True,
-    "nSubiterTurb": 5,
-    "ANKSecondOrdSwitchTol": 1e-4,
-    "ANKCoupledSwitchTol": 1e-6,
-    "ankinnerpreconits": 2,
-    "ankouterpreconits": 2,
-    "anklinresmax": 0.1,
-    # Termination Criteria
-    "L2Convergence": 1e-12,
-    "L2ConvergenceCoarse": 1e-2,
-    "nCycles": 75000,
-}
-
 ################################################################################
 # Default Adflow solver options for AeroStructural problem
 ################################################################################
@@ -127,66 +87,3 @@ default_load_info = {
     'g': [0.0, -9.81, 0.0], # acceleration due to gravity in m/s^2
     'inertial_load_factor': 1.0 # inertial load factor, times of 'g'
 }
-
-################################################################################
-# Default plot options
-################################################################################
-def_plot_options = {
-            'niceplots_style' : 'james-light',
-            'figsize': (20, 6),
-        }
-
-################################################################################
-# Python script to run subprocesses - DO NOT MODIFY
-################################################################################
-python_code_for_subprocess = """
-import argparse
-from mdss.aerostruct import run_problem
-
-parser = argparse.ArgumentParser()
-parser.add_argument("--caseInfoFile", type=str)
-parser.add_argument("--scenarioInfoFile", type=str)
-parser.add_argument("--refLevelDir", type=str)
-parser.add_argument("--aoaList", type=str)
-parser.add_argument("--aeroGrid", type=str)
-parser.add_argument("--structMesh", type=str)
-
-args = parser.parse_args()
-
-run_problem(args.caseInfoFile, args.scenarioInfoFile, args.refLevelDir, args.aoaList, args.aeroGrid, args.structMesh)
-"""
-
-################################################################################
-# Python script to run on a HPC - DO NOT MODIFY
-################################################################################
-python_code_for_hpc = """
-import argparse
-from mdss.run_sim import run_sim
-
-parser = argparse.ArgumentParser()
-parser.add_argument("--inputFile", type=str)
-args = parser.parse_args()
-sim = run_sim(args.inputFile) # Input the simulation info and output dir
-sim.run_problem() # Run the simulation
-"""
-
-################################################################################
-# Template for Greatlakes Job script
-################################################################################
-gl_job_script = """#!/bin/bash
-#SBATCH --job-name={job_name}
-#SBATCH --nodes={nodes}
-#SBATCH --ntasks={nproc}
-#SBATCH --ntasks-per-node=36
-#SBATCH --cpus-per-task=1
-#SBATCH --exclusive
-#SBATCH --mem-per-cpu={mem_per_cpu}
-#SBATCH --time={time}
-#SBATCH --account={account_name}
-#SBATCH --partition=standard
-#SBATCH --mail-type=BEGIN,END,FAIL
-#SBATCH --mail-user={email_id}
-#SBATCH --output={out_file}
-
-python {python_file_path} --inputFile {yaml_file_path}
-"""
