+++ mdss/resources/misc_defaults.py
@@ -1,0 +1,7 @@
+################################################################################
+# Default plot options
+################################################################################
+def_plot_options = {
+            'niceplots_style' : 'james-light',
+            'figsize': (20, 6),
+        }
