+++ mdss/src/__init__.py
@@ -1,0 +1,5 @@
+# src/__init__.py
+from .main import simulation, post_process
+from .aerostruct import Problem
+
+__all__ = ["simulation", "post_process", "Problem"]
