+++ mdss/src/aerostruct.py
@@ -1,9 +1,8 @@
-import os
-import time
-import numpy as np
-import yaml
-import shutil
+# Imports
+import os, time, yaml
+from mpi4py import MPI
 
+# MDO tool imports
 from mphys import MPhysVariables, Multipoint
 from mphys.scenarios import ScenarioAeroStructural, ScenarioAerodynamic
 from mpi4py import MPI
@@ -12,16 +11,16 @@ from baseclasses import AeroProblem
 try: # The following modules are required only for Aerostructural problems
     from tacs.mphys import TacsBuilder
     from funtofem.mphys import MeldBuilder
-    from mdss.tacs_setup import tacs_setup
+    from mdss.src.tacs_config import tacs_setup
+
 except:
     pass
-
-
 import openmdao.api as om
-from mpi4py import MPI
 
-from mdss.helpers import *
-from mdss.templates import *
+from mdss.utils.helpers import ProblemType, load_yaml_file, print_msg, update_om_instance
+from mdss.resources.aero_defaults import default_aero_options_aerodynamic
+from mdss.resources.aerostruct_defaults import *
+
 
 
 comm = MPI.COMM_WORLD
@@ -69,7 +68,6 @@ class Top(Multipoint):
 
     """
     def __init__(self, sim_info):
-        super().__init__()
         self.sim_info = sim_info
         self.problem_type = ProblemType.from_string(sim_info['problem'])  # Convert string to enum
 
@@ -188,141 +186,165 @@ class Top(Multipoint):
 
         # connect to the aero for each scenario
         self.connect("aoa", [variable_to_connect, f"{self.sim_info['scenario_name']}.aero_post.aoa"])
-
-def run_problem(case_info_fpath, scenario_info_fpath, ref_level_dir, aoa_csv_str, aero_grid_fpath, struct_mesh_fpath=None):
-    # Extarct the required info
-    case_info = load_yaml_file(case_info_fpath, comm)
-    scenario_info = load_yaml_file(scenario_info_fpath, comm)
-    aoa_list = [float(x) for x in aoa_csv_str.split(',')]
-    problem_type = ProblemType.from_string(case_info['problem'])  # Convert string to enum
     
-    # Initialize the structrual info dictionaries which remains empty for aerodynamic problems
-    structural_properties = {}
-    load_info = {}
-    solver_options = {}
-    isym=None
+class Problem:
 
-    # Assigning defaults
-    # Read the respective default_aero_options
-    if problem_type == ProblemType.AERODYNAMIC:
-        aero_options = default_aero_options_aerodynamic.copy() # Assign default options for aerodynamic case
-    elif problem_type == ProblemType.AEROSTRUCTURAL:
-        isym = case_info['struct_options']['isym']
-        solver_options = default_solver_options
-        aero_options = default_aero_options_aerostructural.copy() # Assign default aero_options for aerostructural case
-        structural_properties.update(default_struct_properties.copy()) # Assign default structural properties
-        load_info = default_load_info.copy()
-        structural_properties.update({'t': case_info['struct_options']['t']}) # Update thickness with user given values
-        # Update default values with user given data 
-        struct_options = case_info.get('struct_options', {})
-        structural_properties.update(struct_options.get('properties', {}))
-        load_info.update(struct_options.get('load_info', {}))
-        solver_options_updt = struct_options.get('solver_options', {})
-        solver_options['linear_solver_options'].update(solver_options_updt.get('linear_solver_options', {}))
-        solver_options['nonlinear_solver_options'].update(solver_options_updt.get('nonlinear_solver_options', {}))
+    def __init__(self, case_info_fpath, scenario_info_fpath, ref_level_dir, aoa_csv_str, aero_grid_fpath, struct_mesh_fpath=None):
 
-    geometry_info = case_info['geometry_info']
+        # Extarct the required info
+        case_info = load_yaml_file(case_info_fpath, comm)
+        self.case_info = case_info
 
-    # Update aero_options file
-    aero_options.update(case_info.get('aero_options', {}))
-    aero_options['gridFile'] = aero_grid_fpath # Add aero_grid file path to aero_options
+        scenario_info = load_yaml_file(scenario_info_fpath, comm)
+        self.scenario_info = scenario_info
         
-    sim_info = {
-            'problem': case_info['problem'],
-            'scenario_name': scenario_info['name'],
-            'aero_options': aero_options,
-            'chordRef': geometry_info['chordRef'],
-            'areaRef': geometry_info['areaRef'],
-            'scenario_info': scenario_info,
+        aoa_list = [float(x) for x in aoa_csv_str.split(',')]
+        self.aoa_list = aoa_list
+        
+        problem_type = ProblemType.from_string(case_info['problem'])  # Convert string to enum
+        self.problem_type = problem_type
+
+        # Initialize the structrual info dictionaries which remains empty for aerodynamic problems
+        structural_properties = {}
+        load_info = {}
+        solver_options = {}
+        isym=None
+
+        # Assigning defaults
+        # Read the respective default_aero_options
+        if problem_type == ProblemType.AERODYNAMIC:
+            aero_options = default_aero_options_aerodynamic.copy() # Assign default options for aerodynamic case
+        elif problem_type == ProblemType.AEROSTRUCTURAL:
+            isym = case_info['struct_options']['isym']
+            solver_options = default_solver_options
+            aero_options = default_aero_options_aerostructural.copy() # Assign default aero_options for aerostructural case
+            structural_properties.update(default_struct_properties.copy()) # Assign default structural properties
+            load_info = default_load_info.copy()
+            structural_properties.update({'t': case_info['struct_options']['t']}) # Update thickness with user given values
+            # Update default values with user given data 
+            struct_options = case_info.get('struct_options', {})
+            structural_properties.update(struct_options.get('properties', {}))
+            load_info.update(struct_options.get('load_info', {}))
+            solver_options_updt = struct_options.get('solver_options', {})
+            solver_options['linear_solver_options'].update(solver_options_updt.get('linear_solver_options', {}))
+            solver_options['nonlinear_solver_options'].update(solver_options_updt.get('nonlinear_solver_options', {}))
+
+        
+
+        geometry_info = case_info['geometry_info']
+
+        # Update aero_options file
+        aero_options.update(case_info.get('aero_options', {}))
+        aero_options['gridFile'] = aero_grid_fpath # Add aero_grid file path to aero_options
             
-            # Add Structural Info
-            'isym': isym,
-            'tacs_out_dir': ref_level_dir,
-            'struct_mesh_fpath': struct_mesh_fpath,
-            'structural_properties': structural_properties,
-            'load_info': load_info,
-            'solver_options': solver_options,
-        }
-    ################################################################################
-    # OpenMDAO setup
-    ################################################################################
-    os.environ["OPENMDAO_REPORTS"]="0" # Do this to disable report generation by OpenMDAO
-    prob = om.Problem()
-    prob.model = Top(sim_info)
-    prob.setup() # Setup the problem
-    for aoa in aoa_list:
-        prob["aoa"] = float(aoa) # Set Angle of attack
-        aoa_out_dir = os.path.join(ref_level_dir, f"aoa_{aoa}") # name of the aoa output directory
-        aoa_out_dir = os.path.abspath(aoa_out_dir)
-        aoa_info_file = os.path.join(aoa_out_dir, f"aoa_{aoa}.yaml") # name of the simulation info file at the aoa level directory
-        # Update options specific to this aoa
-        prob_updt = update_om_instance(prob, sim_info['scenario_name'], sim_info['problem'])
-        prob_updt.outdir(aoa_out_dir)
-        
-        ################################################################################
-        # Checking for existing sucessful simualtion info
-        ################################################################################ 
-        if os.path.exists(aoa_out_dir):
-            try:
-                with open(aoa_info_file, 'r') as aoa_file:
-                    aoa_sim_info = yaml.safe_load(aoa_file)
-                fail_flag = aoa_sim_info['fail_flag']
-                if fail_flag == 0:
-                    msg = f"Skipping Angle of Attack (AoA): {float(aoa):<5} | Reason: Existing successful simulation found"
-                    print_msg(msg, 'notice', comm)
-                    continue # Continue to next loop if there exists a successful simulation
-            except:
-                fail_flag = 1
-        elif not os.path.exists(aoa_out_dir): # Create the directory if it doesn't exist
-            if comm.rank == 0:
-                os.makedirs(aoa_out_dir)
-        
-        ################################################################################
-        # Run sim when a succesful simulation is not found
-        ################################################################################
-        fail_flag = 0
-        # Run the model
-        aoa_start_time = time.time() # Store the start time
-        try:
-            prob.run_model()
-            fail_flag = 0
-        except:
-            fail_flag = 1
-
-        om.n2(prob, show_browser=False, outfile=os.path.join(aoa_out_dir, "mphys_n2.html")) # Store n2 diagram in the output_directory
-
-        aoa_end_time = time.time() # Store the end time
-        aoa_run_time = aoa_end_time - aoa_start_time # Compute the run time
-
-        prob.model.list_inputs(units=True)
-        prob.model.list_outputs(units=True)
-
-        # Store a Yaml file at this level
-        aoa_out_dic = {
-            # 'refinement_level': refinement_level, # Decide on this later
-            'AOA': float(aoa),
-            'fail_flag': int(fail_flag),
-            'case':case_info['name'],
-            'problem': case_info['problem'],
-            'aero_mesh_fpath': aero_grid_fpath,
-            'scenario_info': scenario_info,
-            'cl': float(prob[f"{sim_info['scenario_name']}.aero_post.cl"][0]),
-            'cd': float(prob[f"{sim_info['scenario_name']}.aero_post.cd"][0]),
-            'wall_time': f"{aoa_run_time:.2f} sec",
-            'out_dir': aoa_out_dir,
-        }
-        try:
-            aoa_out_dic['scenario_info']['exp_data'] = scenario_info['exp_data']
-        except:
-            aoa_out_dic['scenario_info']['exp_data'] = 'Not provided'
-
-        if problem_type == ProblemType.AEROSTRUCTURAL: # Add structural info
-            struct_info_dict = {
+        sim_info = {
+                'aoa_list': aoa_list,
+                'ref_level_dir': ref_level_dir,
+                'problem': case_info['problem'],
+                'scenario_name': scenario_info['name'],
+                'aero_options': aero_options,
+                'chordRef': geometry_info['chordRef'],
+                'areaRef': geometry_info['areaRef'],
+                'scenario_info': scenario_info,
+                
+                # Add Structural Info
+                'isym': isym,
+                'tacs_out_dir': ref_level_dir,
                 'struct_mesh_fpath': struct_mesh_fpath,
                 'structural_properties': structural_properties,
                 'load_info': load_info,
-                'solver_options': solver_options
+                'solver_options': solver_options,
             }
-            aoa_out_dic.update(struct_info_dict)
-        with open(aoa_info_file, 'w') as interim_out_yaml:
-            yaml.dump(aoa_out_dic, interim_out_yaml, sort_keys=False)
+        ################################################################################
+        # OpenMDAO setup
+        ################################################################################
+        os.environ["OPENMDAO_REPORTS"]="0" # Do this to disable report generation by OpenMDAO
+        prob = om.Problem()
+        prob.model = Top(sim_info)
+        prob.setup() # Setup the problem
+
+        self.prob = prob
+        self.sim_info = sim_info
+
+    def run(self):
+
+        for aoa in self.aoa_list:
+            self.prob["aoa"] = float(aoa) # Set Angle of attack
+            aoa_out_dir = os.path.join(self.sim_info.get('ref_level_dir'), f"aoa_{aoa}") # name of the aoa output directory
+            aoa_out_dir = os.path.abspath(aoa_out_dir)
+            aoa_info_file = os.path.join(aoa_out_dir, f"aoa_{aoa}.yaml") # name of the simulation info file at the aoa level directory
+            # Update options specific to this aoa
+            prob_updt = update_om_instance(self.prob, self.sim_info['scenario_name'], self.sim_info['problem'])
+            prob_updt.outdir(aoa_out_dir)
+            
+            ################################################################################
+            # Checking for existing sucessful simualtion info
+            ################################################################################ 
+            if os.path.exists(aoa_out_dir):
+                try:
+                    with open(aoa_info_file, 'r') as aoa_file:
+                        aoa_sim_info = yaml.safe_load(aoa_file)
+                    fail_flag = aoa_sim_info['fail_flag']
+                    if fail_flag == 0:
+                        msg = f"Skipping Angle of Attack (AoA): {float(aoa):<5} | Reason: Existing successful simulation found"
+                        print_msg(msg, 'notice', comm)
+                        continue # Continue to next loop if there exists a successful simulation
+                except:
+                    fail_flag = 1
+            elif not os.path.exists(aoa_out_dir): # Create the directory if it doesn't exist
+                if comm.rank == 0:
+                    os.makedirs(aoa_out_dir)
+            
+            ################################################################################
+            # Run sim when a succesful simulation is not found
+            ################################################################################
+            fail_flag = 0
+            # Run the model
+            aoa_start_time = time.time() # Store the start time
+            try:
+                self.prob.run_model()
+                fail_flag = 0
+            except:
+                fail_flag = 1
+
+            om.n2(self.prob, show_browser=False, outfile=os.path.join(aoa_out_dir, "mphys_n2.html")) # Store n2 diagram in the output_directory
+
+            aoa_end_time = time.time() # Store the end time
+            aoa_run_time = aoa_end_time - aoa_start_time # Compute the run time
+
+            self.prob.model.list_inputs(units=True)
+            self.prob.model.list_outputs(units=True)
+
+            # Store a Yaml file at this level
+            aoa_out_dic = {
+                # 'refinement_level': refinement_level, # Decide on this later
+                'AOA': float(aoa),
+                'fail_flag': int(fail_flag),
+                'case': self.case_info['name'],
+                'problem': self.case_info['problem'],
+                'aero_mesh_fpath': self.sim_info.get('aero_grid_fpath'),
+                'scenario_info': self.scenario_info,
+                'cl': float(self.prob[f"{self.sim_info['scenario_name']}.aero_post.cl"][0]),
+                'cd': float(self.prob[f"{self.sim_info['scenario_name']}.aero_post.cd"][0]),
+                'wall_time': f"{aoa_run_time:.2f} sec",
+                'out_dir': aoa_out_dir,
+            }
+            try:
+                aoa_out_dic['scenario_info']['exp_data'] = self.scenario_info['exp_data']
+            except:
+                aoa_out_dic['scenario_info']['exp_data'] = 'Not provided'
+
+            if self.problem_type == ProblemType.AEROSTRUCTURAL: # Add structural info
+                struct_info_dict = {
+                    'struct_mesh_fpath': self.sim_info['struct_mesh_fpath'],
+                    'structural_properties': self.sim_info['structural_properties'],
+                    'load_info': self.sim_info['load_info'],
+                    'solver_options': self.sim_info['solver_options'],
+                }
+                aoa_out_dic.update(struct_info_dict)
+            with open(aoa_info_file, 'w') as interim_out_yaml:
+                yaml.dump(aoa_out_dic, interim_out_yaml, sort_keys=False)
+
+
+
+
