+++ mdss/src/main.py
@@ -13,24 +13,22 @@ from matplotlib.legend import Legend
 import niceplots
 from mpi4py import MPI
 
-from mdss.helpers import *
-from mdss.templates import def_plot_options
-from mdss.yaml_config import check_input_yaml, ref_plot_options
-from mdss.aerostruct import run_problem as run_aerostructural
+from mdss.utils.helpers import load_yaml_file, load_csv_data, make_dir, print_msg, MachineType
+from mdss.src.main_helper import execute, submit_job_on_hpc
+from mdss.resources.misc_defaults import def_plot_options
+from mdss.resources.yaml_config import ref_plot_options, check_input_yaml
+
 
 comm = MPI.COMM_WORLD
 
-class run_sim():
+class simulation():
     """
-    Executes ADflow simulations using the `Top` class defined in [`aerostruct.py`](aerostruct.py).
+    Executes aero(structural) simulations using the `Top` class defined in [`aerostruct.py`](aerostruct.py).
 
-    This class sets up, runs, and post-processes aerodynamic and/or aerostructural simulations based on input parameters provided via a YAML configuration file. It validates the input, manages directories, and handles outputs, including plots and summary files.
+    This class sets up and runs aerodynamic and/or aerostructural simulations based on input parameters provided via a YAML configuration file. It validates the input, manages directories, and handles outputs, including summary files. The simulations are run using subprocesses.
 
     Methods
     -------
-    **run_problem()**
-        Sets up and runs the OpenMDAO problem for all hierarchies, cases, and refinement levels in subprocesses.
-
     **run()**
         Helps to execute the simulation on either a local machine or an HPC.
 
@@ -50,214 +48,14 @@ class run_sim():
         self.sim_info = load_yaml_file(self.info_file, comm)
         self.out_dir = os.path.abspath(self.sim_info['out_dir'])
         self.machine_type = MachineType.from_string(self.sim_info['machine_type'])  # Convert string to enum
-
         # Additional options
         self.final_out_file = os.path.join(self.out_dir, "overall_sim_info.yaml") # Set the overall simulation info file name.
         self.subprocess_flag = True # To toggle opting subprocess.
-        self.record_subprocess = False # To toggle to recored subprocess output.
+        self.record_subprocess = False # To toggle to record subprocess output.
 
         # Create the output directory if it doesn't exist
         make_dir(self.out_dir, comm)
-    
-    ################################################################################
-    # Code for running simulations
-    ################################################################################   
-    def run_problem(self):
-        """
-        Runs the aerodynamic and/or aerostructural simulations as subprocesses.
 
-        This method iterates through all hierarchies, cases, refinement levels, and angles of attack defined in the input YAML file. Runs the simulation by calling `aerostruct.py`, and stores the results.
-
-        Outputs
-        -------
-        - **A CSV file**:
-            Contains results for each angle of attack at the current refinement level.
-        - **A YAML file**:
-            Stores simulation data for each angle of attack in the corresponding directory.
-        - **A final YAML file**:
-            Summarizes all simulation results across hierarchies, cases, and refinement levels.
-
-        Notes
-        -----
-        This method ensures that:
-
-        - Existing successful simulations are skipped.
-        - Directories are created dynamically if they do not exist.
-        - Simulation results are saved in structured output files.
-        """
-
-        # Store a copy of input YAML file in output directory
-        input_yaml_file = os.path.join(self.out_dir, "input_file.yaml")
-        if comm.rank == 0:
-            with open(input_yaml_file, 'w') as input_yaml_handle:
-                yaml.dump(self.sim_info, input_yaml_handle, sort_keys=False)
-        
-        sim_info_copy = copy.deepcopy(self.sim_info) # Copying to run the loop
-        sim_out_info = copy.deepcopy(self.sim_info) # Copying to write the output YAML file
-        start_time = time.time()
-        start_wall_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-
-        for hierarchy, hierarchy_info in enumerate(sim_info_copy['hierarchies']): # loop for Hierarchy level
-
-            for case, case_info in enumerate(hierarchy_info['cases']): # loop for cases in hierarchy
-                # Assign problem type
-                problem_type = ProblemType.from_string(case_info['problem'])  # Convert string to enum
-
-                # Define case level outptut directory
-                case_out_dir = os.path.join(self.out_dir, hierarchy_info['name'], case_info['name'])
-                make_dir(case_out_dir, comm)
-                comm.Barrier()
-                # Save case info yaml file in the case_out_dir to pass in subprocess
-                case_info_fpath = os.path.join(case_out_dir, "case_info.yaml")
-                with open(case_info_fpath, 'w') as case_info_fhandle:
-                    yaml.dump(case_info, case_info_fhandle, sort_keys=False)
-
-                for scenario, scenario_info in enumerate(case_info['scenarios']): # loop for scenarios that may present
-                    
-                    # Define scenario level output directory
-                    scenario_out_dir = os.path.join(case_out_dir, scenario_info['name'])
-                    make_dir(scenario_out_dir, comm)
-                    comm.Barrier()
-                    # Save case info yaml file in the case_out_dir to pass in subprocess
-                    scenario_info_fpath = os.path.join(scenario_out_dir, "scenario_info.yaml")
-                    with open(scenario_info_fpath, 'w') as scenario_info_fhandle:
-                        yaml.dump(scenario_info, scenario_info_fhandle, sort_keys=False)
-                    
-                    # Extract the Angle of attacks for which the simulation has to be run
-                    aoa_list = scenario_info['aoa_list']
-                    aoa_csv_string = ",".join(map(str, [float(aoa) for aoa in aoa_list]))
-                    scenario_sim_info = {} # Creating scenario level sim info dictionary for overall sim info file
-
-                    for ii, mesh_file in enumerate(case_info['mesh_files']): # Loop for refinement levels
-                        # Print simulation info message
-                        msg = f"{'Hierarchy':<20}: {hierarchy_info['name']}\n{'Case Name':<20}: {case_info['name']}\n{'Scenario':<20}: {scenario_info['name']}\n{'Aero Mesh File':<20}: {mesh_file}"
-                        print_msg(msg, f"{'SIMULATION INFO':^30}", comm)
-                        AOAList = []
-                        CLList = []
-                        CDList = []
-                        TList = []
-                        FList = [] # Fail flag list
-
-                        refinement_level_dict = {} # Creating refinement level sim info dictionary for overall sim info file
-                        refinement_out_dir = os.path.join(scenario_out_dir, f"{mesh_file}")
-                        make_dir(refinement_out_dir, comm)
-                        aero_grid_fpath = os.path.join(case_info['meshes_folder_path'], mesh_file)
-                        # Add struct mesh file for aerostructural case else set it to none
-                        if problem_type == ProblemType.AEROSTRUCTURAL:
-                            struct_mesh_file = case_info['struct_options']['mesh_fpath']
-                        else:
-                            struct_mesh_file = 'none'
-
-                        # Run subprocess
-                        # Initially running all the aoa in a subprocess. However the optimal number of aoa for single subprocess should be determined and modified accordingly.
-                        other_sim_info = {key: value for key, value in sim_info_copy.items() if key != 'hierarchies'} # To pass just the sim_info without hierarchies
-                        if self.subprocess_flag is True:
-                            run_as_subprocess(other_sim_info, case_info_fpath, scenario_info_fpath, refinement_out_dir, aoa_csv_string, aero_grid_fpath, struct_mesh_file,  comm, self.record_subprocess)
-                        elif self.subprocess_flag is False:
-                            run_aerostructural(case_info_fpath, scenario_info_fpath, refinement_out_dir, aoa_csv_string, aero_grid_fpath, struct_mesh_file)
-                        failed_aoa_list = [] # Initiate a list to store a list of aoa failed in this refinement level
-                
-                        for aoa in aoa_list: # loop for angles of attack reads the info, adds additional info if needed for the output file for each aoa
-                            aoa = float(aoa) # making sure aoa is a float
-                            aoa_out_dir = os.path.join(refinement_out_dir, f"aoa_{aoa}") # aoa output directory -- Written to store in the parent directory
-                            aoa_info_file = os.path.join(aoa_out_dir ,f"aoa_{aoa}.yaml") # name of the simulation info file at the aoa level directory
-                            aoa_level_dict = {} # Creating aoa level sim info dictionary for overall sim info file
-
-                            # Checking for existing sucessful simualtion info, 
-                            try:
-                                with open(aoa_info_file, 'r') as aoa_file: # open the simulation info file
-                                    aoa_sim_info = yaml.safe_load(aoa_file)
-
-                                fail_flag = aoa_sim_info['fail_flag'] # Read the fail flag
-
-                                if fail_flag == 0: # Refers successful simulation and makes sure only the sucessful simulations are added to the csv file.
-                                    # Add the simulation info to list to be saved as a csv file in the refinement out directory
-                                    AOAList.append(aoa_sim_info['AOA'])
-                                    CLList.append(aoa_sim_info['cl'])
-                                    CDList.append(aoa_sim_info['cd'])
-                                    TList.append(float(aoa_sim_info['wall_time'].replace(" sec", "")))
-                                    FList.append(fail_flag)
-
-                                    # Store the basic info that is needed to be stored in refinement level dictionary
-                                    aoa_level_dict = {
-                                        'cl': float(aoa_sim_info['cl']),
-                                        'cd': float(aoa_sim_info['cd']),
-                                        'wall_time': aoa_sim_info['wall_time'],
-                                        'fail_flag': int(fail_flag),
-                                        'out_dir': aoa_out_dir,
-                                    }
-                                    refinement_level_dict["failed_aoa"] = failed_aoa_list
-                                    refinement_level_dict[f"aoa_{aoa}"] = aoa_level_dict
-                                    
-                                elif fail_flag == 1: # refers to failed simulation
-                                    failed_aoa_list.append(aoa) # Add to the list of failed aoa
-                                
-                                # Save the aoa_out_dict as an yaml file with the updated info
-                                with open(aoa_info_file, 'w') as interim_out_yaml:
-                                    yaml.dump(aoa_sim_info, interim_out_yaml, sort_keys=False)
-
-                            except:
-                                failed_aoa_list.append(aoa) # Add to the list of failed aoa
-                        ################################# End of AOA loop ########################################
-
-                        # Write simulation results to a csv file
-                        refinement_level_data = {
-                            "Alpha": [f"{alpha:6.2f}" for alpha in AOAList],
-                            "CL": [f"{cl:8.4f}" for cl in CLList],
-                            "CD": [f"{cd:8.4f}" for cd in CDList],
-                            "FFlag": [f"{int(FF):12f}" for FF in FList],
-                            "WTime": [f"{wall_time:10.2f}" for wall_time in TList]
-                        }
-
-                        # Define the output file path
-                        refinement_level_dir = os.path.dirname(aoa_out_dir)
-                        ADflow_out_file = os.path.join(refinement_level_dir, "ADflow_output.csv")
-                        
-                        df = pd.DataFrame(refinement_level_data) # Create a panda DataFrame
-                        df.to_csv(ADflow_out_file, index=False)# Write the DataFrame to a CSV file
-
-                        # Add csv file location to the overall simulation out file
-                        refinement_level_dict['csv_file'] = ADflow_out_file
-                        refinement_level_dict['refinement_out_dir'] = refinement_level_dir
-
-                        # Add refinement level dict to scenario level dict
-                        scenario_sim_info[f"{mesh_file}"] = refinement_level_dict
-                    ################################# End of refinement loop ########################################
-
-                    # Add scenario level simulation to the overall simulation out file
-                    scenario_out_dir = os.path.dirname(refinement_level_dir)
-                    scenario_sim_info['scenario_out_dir'] = scenario_out_dir
-                    sim_out_info['hierarchies'][hierarchy]['cases'][case]['scenarios'][scenario]['sim_info'] = scenario_sim_info
-
-                    if os.path.exists(scenario_info_fpath): # Remove the scenario_info yaml file
-                        if comm.rank==0:
-                            os.remove(scenario_info_fpath)
-                ################################# End of scenarios loop ########################################
-                
-                if os.path.exists(case_info_fpath): # Remove the case_info yaml file
-                    if comm.rank==0:
-                        os.remove(case_info_fpath)
-            ################################# End of case loop ########################################
-        ################################# End of hierarchy loop ########################################
-
-        end_time = time.time()
-        end_wall_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        net_run_time = end_time - start_time
-        sim_out_info['overall_sim_info'] = {
-            'start_time': start_wall_time,
-            'end_time': end_wall_time,
-            'total_wall_time': f"{net_run_time:.2f} sec"
-        }
-
-        # Store the final simulation out file.
-        if os.path.exists(self.final_out_file):
-            prev_sim_info = load_yaml_file(self.final_out_file, comm) # Load the previous sim_out_info
-            deep_update(sim_out_info, prev_sim_info)
-
-        if comm.rank == 0:
-            with open(self.final_out_file, 'w') as final_out_yaml_handle:
-                yaml.dump(sim_out_info, final_out_yaml_handle, sort_keys=False)
-        comm.Barrier()
     
     ################################################################################
     # Code for user to run simulations
@@ -275,17 +73,18 @@ class run_sim():
         """
         sim_info_copy = copy.deepcopy(self.sim_info)
         if self.machine_type == MachineType.LOCAL: # Running on a local machine
-            self.run_problem()
+            execute(self)
 
         elif self.machine_type == MachineType.HPC: # Running on a HPC currently supports Great Lakes.
             submit_job_on_hpc(sim_info_copy, self.info_file, comm) # Submit job script
+
                        
 ################################################################################
 # Code for Post Processing
 ################################################################################
 class post_process:
     """
-    Performs post-processing operations for ADflow simulation results.
+    Performs post-processing operations for simulation results.
 
     This class provides functionality to visualize and compare aerodynamic performance data
     such as Lift Coefficient ($C_L$) and Drag Coefficient ($C_D$) against Angle of Attack (Alpha),
@@ -316,6 +115,7 @@ class post_process:
     - *_get_marker_style(idx)*  
         Returns a marker style based on the index, used to distinguish between scenarios visually.
     """
+
     def __init__(self, out_dir: str, plot_options: dict={}):
         self.out_dir = os.path.abspath(out_dir)
         self.final_out_file = os.path.join(self.out_dir, "overall_sim_info.yaml") # Setting the overall simulation info file.
