+++ output/2d_clean/NACA0012/cruise_1/naca0012_L0.cgns/ADflow_output.csv
@@ -1,0 +1,1 @@
+Alpha,CL,CD,FFlag,WTime
