+++ output/input_file.yaml
@@ -6,7 +6,7 @@ hierarchies:
   cases:
   - name: NACA0012
     problem: Aero
-    meshes_folder_path: ../grids
+    meshes_folder_path: /home/sanjan/PhD/mdss/examples/grids
     mesh_files:
     - naca0012_L0.cgns
     geometry_info:
@@ -24,9 +24,9 @@ hierarchies:
     scenarios:
     - name: cruise_1
       aoa_list:
-        - 0
-        - 5
+      - 0
+      - 5
       reynolds: 3900000.0
       mach: 0.3
       T: 298.0
-      exp_data: ./exp_data/naca0012.csv
+      exp_data: /home/sanjan/PhD/mdss/examples/exp_data/naca0012.csv
