+++ output/overall_sim_info.yaml
@@ -1,0 +1,41 @@
+out_dir: output
+machine_type: local
+nproc: 8
+hierarchies:
+- name: 2d_clean
+  cases:
+  - name: NACA0012
+    problem: Aero
+    meshes_folder_path: /home/sanjan/PhD/mdss/examples/grids
+    mesh_files:
+    - naca0012_L0.cgns
+    geometry_info:
+      chordRef: 1.0
+      areaRef: 1.0
+    aero_options:
+      useANKSolver: true
+      nSubiterTurb: 20
+      useNKSolver: false
+      NKSwitchTol: 1.0e-06
+      ANKCoupledSwitchTol: 0.001
+      ANKSecondOrdSwitchTol: 1.0e-12
+      L2Convergence: 1.0e-08
+      nCycles: 150000
+    scenarios:
+    - name: cruise_1
+      aoa_list:
+      - 0
+      - 5
+      reynolds: 3900000.0
+      mach: 0.3
+      T: 298.0
+      exp_data: /home/sanjan/PhD/mdss/examples/exp_data/naca0012.csv
+      sim_info:
+        naca0012_L0.cgns:
+          csv_file: /home/sanjan/PhD/mdss/output/2d_clean/NACA0012/cruise_1/naca0012_L0.cgns/ADflow_output.csv
+          refinement_out_dir: /home/sanjan/PhD/mdss/output/2d_clean/NACA0012/cruise_1/naca0012_L0.cgns
+        scenario_out_dir: /home/sanjan/PhD/mdss/output/2d_clean/NACA0012/cruise_1
+overall_sim_info:
+  start_time: '2025-03-31 18:21:00'
+  end_time: '2025-03-31 18:22:05'
+  total_wall_time: 68.64 sec
